shared supabase client in new clients folder, refactored badges.ts to remove interface types to new types folder and file
