+++ services/api/src/routes/badges.ts
@@ -1,71 +1,8 @@
 import { Router, Request, Response } from "express";
-import { createClient } from "@supabase/supabase-js";
-/*
-import type {
-  Badge,
-  UserBadge,
-  BadgeWithEarnedStatus,
-  AwardBadgeRequest,
-  AwardBadgeResponse,
-} from "../../../packages/types/src/index.ts";
-*/
-
-export interface Badge {
-  id: string;
-  code: string;
-  name: string;
-  description: string | null;
-  icon: string | null;
-  criteria: BadgeCriteria;
-  created_at: string;
-}
-
-export interface BadgeCriteria {
-  type: "workout_count" | "streak" | "morning_workout" | "workout_type";
-  target: number;
-  workout_type?: "strength" | "cardio" | "yoga" | "flexibility";
-}
-
-export interface UserBadge {
-  id: string;
-  user_id: string;
-  badge_id: string;
-  awarded_at: string;
-  badge?: Badge;
-}
-
-export interface Streak {
-  id: string;
-  user_id: string;
-  current_days: number;
-  longest_days: number;
-  last_active_date: string | null;
-  updated_at: string;
-}
-
-export interface BadgeWithEarnedStatus extends Badge {
-  earned: boolean;
-  earned_at?: string;
-}
-
-export interface AwardBadgeRequest {
-  user_id: string;
-  badge_code: string;
-}
-
-export interface AwardBadgeResponse {
-  success: boolean;
-  badge?: UserBadge;
-  message?: string;
-}
+import { supabase } from "../clients/supabaseClient";
+import type { Badge, BadgeWithEarnedStatus, AwardBadgeRequest, AwardBadgeResponse } from "../types/badges";
 
 const router: Router = Router();
-
-// Initialize Supabase client
-const supabase = createClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.SUPABASE_SERVICE_ROLE_KEY!
-);
 
 //get badges
 router.get("/", async (req: Request, res: Response) => {
